server/requirements.txt
quart
dnspython
base64
